+++ src/rtextures.c
@@ -1250,7 +1250,7 @@ Image ImageText(const char *text, int fontSize, Color color)
 {
     Image imText = { 0 };
 #if defined(SUPPORT_MODULE_RTEXT)
-    int defaultFontSize = 10;   // Default Font chars height in pixel
+    int defaultFontSize = 10;   // Default RayFont chars height in pixel
     if (fontSize < defaultFontSize) fontSize = defaultFontSize;
     int spacing = fontSize/defaultFontSize;
     imText = ImageTextEx(GetFontDefault(), text, (float)fontSize, (float)spacing, color);   // WARNING: Module required: rtext
@@ -1262,7 +1262,7 @@ Image ImageText(const char *text, int fontSize, Color color)
 }
 
 // Create an image from text (custom sprite font)
-Image ImageTextEx(Font font, const char *text, float fontSize, float spacing, Color tint)
+Image ImageTextEx(RayFont font, const char *text, float fontSize, float spacing, Color tint)
 {
     Image imText = { 0 };
 #if defined(SUPPORT_MODULE_RTEXT)
@@ -3155,7 +3155,7 @@ void ImageDrawText(Image *dst, const char *text, int posX, int posY, int fontSiz
 }
 
 // Draw text (custom sprite font) within an image (destination)
-void ImageDrawTextEx(Image *dst, Font font, const char *text, Vector2 position, float fontSize, float spacing, Color tint)
+void ImageDrawTextEx(Image *dst, RayFont font, const char *text, Vector2 position, float fontSize, float spacing, Color tint)
 {
     Image imText = ImageTextEx(font, text, fontSize, spacing, tint);